Read More Alef Protocol
Alef Protocol
This type of encryption is different from other encryptions
Unlike other encryption protocols, a private key, such as symmetric encryption, is used in encryption and decryption.
But the difference between this type of encryption and other text encryption is that unlike them, the key is created by the algorithm and the same key is used for decryption (the key is somehow placed in the hash code) in the password. The user gives the algorithm a broken protocol or key to the algorithm, and based on the key, the hash code is encrypted, which is usually not possible without decrypting the hashed password.

Note: Unlocking the hash password is not possible without a key. Dear users, be diligent in selecting and maintaining the key.
The next difference in this encryption compared to other encryptions is the type of password output (no need to explain)

Attention:
In this type of encryption, only the use of ski characters is allowed
This type of encryption can be done in secret chats (using a shared private key).

Encryption of database passwords and confidential information and ...
used
In the attached file, you can see the  C# source code example of working with this encryption, as well as the broken key program with Alef Protocol.